add --font-size cmdline option
Allows TextRenderer to provide a FontOptions type which will house
any/all options for it's fonts.  It provides an eql function and a
way to parse the --font-size parameter.  The CellType renderer will
parse not only the raw hieght (before DPI scaling) but also parse an
optional aspect ratio from the size.